Split testcases
